requirements.txt
# Core dependencies
requests==2.31.0
PyGithub==1.59.1
python-dateutil==2.8.2
python-dotenv==1.0.0
openai==0.28.1

# Development dependencies
pytest==7.4.0
black==23.7.0
flake8==6.0.0 